Rename folder and fix tests
