+++ test/clients/actorSyncClient.test.ts
@@ -2,7 +2,7 @@ import { IModLoaderAPI } from 'modloader64_api/IModLoaderAPI'
 import { IOOTCore, IOotHelper } from 'modloader64_api/OOT/OOTAPI'
 import { ActorHealthSyncController } from '../../src/ml64-npc-sync/controllers/actorHealthSync/actorHealthSyncController'
 import { ActorSyncClient } from '../../src/ml64-npc-sync/clients/ActorSyncClient'
-import { ActorPositionSyncController } from '../../src/ml64-npc-sync/controllers/actorPositionSyncController.ts/actorPositionSyncController'
+import { ActorPositionSyncController } from '../../src/ml64-npc-sync/controllers/actorPositionSyncController/actorPositionSyncController'
 import { INetworkController } from '../../src/ml64-npc-sync/controllers/interfaces/networkController'
 import { ActorHealthSyncPacket, ACTOR_HEALTH_SYNC_PACKET_TAG } from '../../src/ml64-npc-sync/packets/actorHealthSyncPacket'
 import { ActorHealthData } from '../../src/ml64-npc-sync/model/actorHealthData'
