+++ tsconfig.json
@@ -31,7 +31,7 @@
     "tools/**/*.ts",
     "src/modloader/consoles/mupen/resources/test.js",
     "API/src/IAnalyticsManager.ts"
-  ],
+, "src/ml64-npc-sync/controllers/actorPositionSyncController"  ],
   "exclude": [
     "node_modules",
     "API",
